Выполнил задание по теме: Приведение и преобразование типов

+++ Course/Models/Employee.cs
@@ -3,9 +3,9 @@
 public class Employee : Person
 {
     public string Contract { get; private set; }
-    public int Salary { get; private set; }
+    public double Salary { get; private set; }
 
-    public Employee(string firstName, string lastName, string contract, int salary) : base(firstName, lastName)
+    public Employee(string firstName, string lastName, string contract, double salary) : base(firstName, lastName)
     {
         SetContract(contract);
         SetSalary(salary);
@@ -21,7 +21,7 @@ public class Employee : Person
         Contract = contract;
     }
 
-    public void SetSalary(int salary)
+    public void SetSalary(double salary)
     {
         if (salary < 0)
         {
